Write node coordinates in interleaved format.
Using HDF5 chunked IO to enable the on-disk layout of node coordinates to 
be column-major results in deadlocks when writing in parallel on cosmea.
So this is plan B: just write the node coorindates in interleaved format.
This requires extra buffering of node coordinate data when writing, but 
will hopefully avoid the long read times observed on cosmea.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@3957 6162379e-bd28-0410-9a7d-b7f4dcfcad3c